requirements.txt
flask
requests
beautifulsoup4
dnspython
fpdf
flask_sqlalchemy
flask_login
werkzeug
email_validator
